add case insensitive address checks and move logic to cleaners file
