+++ src/data/positions.js
@@ -1,0 +1,33 @@
+const positions = {
+  lefts: [
+    'CHEMOTREAT (true)',
+    'IMMUNOTREAT (true)',
+    'HORMONETREATMENT (true)',
+    'RADIOTREAT (true)',
+    'SURGERY (true)',
+    'HIV(true)',
+    'SPECIFICALLYPOSITIVE (true)',
+    'OTHERTREATMENT(TRUE)',
+    'BAS (true)',
+    'TELEPHONENO(+254)',
+    'OTHERCONCURRENTILLNESS(true)',
+  ],
+  rights: [
+    'MOR(Matching NCI Codes)',
+    'TOP(Matching NCI Codes)',
+    'ADDR (Sub County)',
+    'WARD',
+    'NOKNUMBER',
+    'NOKNAME',
+    'ADDR (County)',
+  ],
+  yes: [
+    'CHEMOTREAT (true)',
+    'IMMUNOTREAT (true)',
+    'HORMONETREATMENT (true)',
+    'RADIOTREAT (true)',
+    'SURGERY (true)',
+  ],
+};
+
+export default positions;
