+++ src/helpers/transform.js
@@ -1,5 +1,5 @@
 const isInvalid = (input) => input
-&& !/UNKNOWN|Unknown|NONE|N\/A|Invalid code|Invalid category.+/g.test(input);
+  && !/UNKNOWN|Unknown|NONE|N\/A|Invalid code|Invalid category.+/g.test(input);
 
 const replaceAddr = (addr, txt = '') => `${addr
   .toString()
@@ -21,26 +21,29 @@ const insertCol = (title, col, right = false, yes = null, data) => {
     const filledCols = data.map((row, i) => {
       if (i === 0) {
         return row;
-      } if (
+      }
+      if (
         title.toString().toLowerCase().includes('hiv')
-          || title.toString().toUpperCase().includes('SPECIFICALLYPOSITIVE')
+        || title.toString().toUpperCase().includes('SPECIFICALLYPOSITIVE')
       ) {
         row.splice(
           right ? checked + 1 : checked,
           0,
 
-          !!(row[checked]
-              && (row[checked].toString().toLowerCase().includes('positive')
-                || row[checked].toString().toLowerCase() === 'yes')),
+          !!(
+            row[checked]
+            && (row[checked].toString().toLowerCase().includes('positive')
+              || row[checked].toString().toLowerCase() === 'yes')
+          ),
         );
       } else if (col.includes('TELEPHONENO(+254)')) {
         row.splice(
           checked,
           0,
           row[checked]
-              && !/UNKNOWN|Unknown|County|NONE|N\/A|Invalid code+/g.test(
-                row[checked],
-              )
+            && !/UNKNOWN|Unknown|County|NONE|N\/A|Invalid code+/g.test(
+              row[checked],
+            )
             ? `+254${row[checked].toString().replace(/[^0-9]/g, '')}`
             : '',
         );
@@ -49,7 +52,7 @@ const insertCol = (title, col, right = false, yes = null, data) => {
           checked + 1,
           0,
           row[checked]
-              && !/UNKNOWN|Unknown|NONE|N\/A|Invalid code+/g.test(row[checked])
+            && !/UNKNOWN|Unknown|NONE|N\/A|Invalid code+/g.test(row[checked])
             ? `${row[checked]
               .toString()
               .replace(/[^A-Za-z ]/g, '')
@@ -74,9 +77,7 @@ const insertCol = (title, col, right = false, yes = null, data) => {
         row.splice(
           checked + 1,
           0,
-          isInvalid(row[checked])
-            ? replaceAddr(row[checked], 'Sub County')
-            : '',
+          isInvalid(row[checked]) ? replaceAddr(row[checked], 'Sub County') : '',
         );
       } else if (col.includes('WARD')) {
         row.splice(
@@ -89,26 +90,25 @@ const insertCol = (title, col, right = false, yes = null, data) => {
           right ? checked + 1 : checked,
           0,
           right
-            ? !!(row[checked] && row[checked].toString().toLowerCase() === 'yes')
-            : !!(row[checked] && row[checked].toString().toLowerCase() === 'yes'),
+            ? !!(
+              row[checked] && row[checked].toString().toLowerCase() === 'yes'
+            )
+            : !!(
+              row[checked] && row[checked].toString().toLowerCase() === 'yes'
+            ),
         );
       } else {
         row.splice(
           right ? checked + 1 : checked,
           0,
-          right ? (!!row[checked]) : !!row[checked],
+          right ? !!row[checked] : !!row[checked],
         );
       }
 
       return row;
     });
-
-    // setData(filledCols);
-
-    // setVisible(false);
     return filledCols;
   }
-  //   message.error(`Column ${col} already exists`);
   return false;
 };
